+++ lib/Elastica/Bulk/Action/IndexDocument.php
@@ -30,14 +30,14 @@ class IndexDocument extends AbstractDocument
     protected function _getMetadata(AbstractUpdateAction $action): array
     {
         return $action->getOptions([
-            'index',
-            'type',
-            'id',
+            '_index',
+            '_type',
+            '_id',
             'version',
             'version_type',
             'routing',
             'parent',
             'retry_on_conflict',
-        ], true);
+        ]);
     }
 }