+++ test/Elastica/DocumentTest.php
@@ -151,33 +151,30 @@ class DocumentTest extends BaseTest
         $document->setParent('2');
         $document->setId(1);
 
-        $options = $document->getOptions(['index', 'type', 'id', 'parent']);
+        $options = $document->getOptions(['_index', '_type', '_id', 'parent']);
 
         $this->assertInternalType('array', $options);
         $this->assertCount(3, $options);
-        $this->assertArrayHasKey('index', $options);
-        $this->assertArrayHasKey('id', $options);
+        $this->assertArrayHasKey('_index', $options);
+        $this->assertArrayHasKey('_id', $options);
         $this->assertArrayHasKey('parent', $options);
-        $this->assertEquals('index', $options['index']);
-        $this->assertEquals(1, $options['id']);
+        $this->assertEquals('index', $options['_index']);
+        $this->assertEquals(1, $options['_id']);
         $this->assertEquals('2', $options['parent']);
-        $this->assertArrayNotHasKey('type', $options);
+        $this->assertArrayNotHasKey('_type', $options);
         $this->assertArrayNotHasKey('op_type', $options);
-        $this->assertArrayNotHasKey('_index', $options);
-        $this->assertArrayNotHasKey('_id', $options);
-        $this->assertArrayNotHasKey('_parent', $options);
+        $this->assertArrayNotHasKey('index', $options);
+        $this->assertArrayNotHasKey('id', $options);
 
-        $options = $document->getOptions(['parent', 'op_type', 'percolate'], true);
+        $options = $document->getOptions(['parent', 'op_type', 'percolate']);
 
         $this->assertInternalType('array', $options);
         $this->assertCount(2, $options);
-        $this->assertArrayHasKey('_parent', $options);
-        $this->assertArrayHasKey('_op_type', $options);
-        $this->assertEquals('2', $options['_parent']);
-        $this->assertEquals('create', $options['_op_type']);
+        $this->assertArrayHasKey('parent', $options);
+        $this->assertArrayHasKey('op_type', $options);
+        $this->assertEquals('2', $options['parent']);
+        $this->assertEquals('create', $options['op_type']);
         $this->assertArrayNotHasKey('percolate', $options);
-        $this->assertArrayNotHasKey('op_type', $options);
-        $this->assertArrayNotHasKey('parent', $options);
     }
 
     /**